Update README.md
In this , we will analyze bike sharing data from Capital Bikeshare in 
Washington, D.C., USA, for the period between January 1, 2011, and December 31, 
2012. The data is aggregated on an hourly basis. This means that no initial and final 
locations of the individual rides are available, but only the total number of rides per 
hour. Nevertheless, additional meteorological information is available in the data, 
which could serve as a driving factor for identifying the total number of requests for a 
specific time frame (bad weather conditions could have a substantial impact on bike 
sharing demand).